Adds GitHub links to projects
I believe that adding the github links is a good idea because it shows
that our projects are open source. I realize that if people go to our
github organization, they can see most of our projects, but I feel that
them being right on the website, albeit hidden in the title of the
project. This commit also adds a link to audiophiler in the Harold
project.

changes the EventLCD link to the CSH repo instead of Drews