Remove max-height styling for slick carousel, add scroll down button
Removes the max-height styling for the scrolling carousel.

Adds the carousel scroll down button, which solves the design issue
of images taking up the whole page (which upon first viewing might
lead someone to believe that is all the page contains) by providing
an indicator+button that scrolls to reveal the rest of the page.

Upgrades jquery from slim to full. This is required so we can use
animate for the smooth scrolling functionality.

Moves all javascript libraries from the footer to header, if this
is undesirable for any particular reason (I really didn't see any
issues), just say so and I'll pull it from the PR.

Resolves #280